# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quiz/sessions:
    post:
      operationId: startQuizSession
      tags: [quiz]
      summary: Start a server-scored quiz session
      description: |
        Creates a quiz session for an owned material. Returns questions without
        correctIndex. The server stores the answer key internally.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartQuizSessionRequest"
      responses:
        "201":
          description: Quiz session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSessionResponse"
        "400":
          description: Validation error (invalid materialId or totalQuestions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Material not owned by current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Material not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: >
            One or more selected materials are not ready for study (unsupported
            format, extraction failed, or not enough extractable text).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /study/flashcards:
    post:
      operationId: generateFlashcards
      tags: [quiz]
      summary: Generate content-grounded flashcards from owned, ready materials
      description: |
        Verifies all selected materials are owned and extractionStatus=ready with
        meaningful extracted text, then returns flashcards grounded in the actual
        content. Generated separately from quiz answer keys. No filename fallback.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardsRequest"
      responses:
        "200":
          description: Flashcards generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardsResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: One or more materials not found or not owned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: One or more materials not ready for study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /quiz/sessions/{sessionId}/complete:
    post:
      operationId: completeQuizSession
      tags: [quiz]
      summary: Complete a quiz session and award coins (idempotent)
      description: |
        Accepts the user's answers array, scores against the server answer key,
        inserts one attempt record, awards wallet coins, and marks session complete.
        Idempotent: same idempotencyKey on a completed session returns the existing
        attempt without re-awarding coins. Different key on completed session
        returns existing attempt without re-awarding.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteQuizSessionRequest"
      responses:
        "200":
          description: Session completed (or already completed — returns existing attempt)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttempt"
        "400":
          description: Validation error (wrong answer array length, expired session)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Session not owned by current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Session already completed but attempt record unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /quiz/attempts:
    get:
      operationId: listQuizAttempts
      tags: [quiz]
      summary: List quiz attempts for the current user
      responses:
        "200":
          description: List of attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizAttempt"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    QuizQuestion:
      type: object
      required: [question, options]
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
          minItems: 4
          maxItems: 4
    StartQuizSessionRequest:
      type: object
      required: [materialIds, totalQuestions]
      properties:
        materialIds:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 20
          description: Owned material ids to combine into one unified quiz.
        totalQuestions:
          type: integer
          enum: [10, 20, 30]
    QuizSessionResponse:
      type: object
      required: [sessionId, questions, expiresAt]
      properties:
        sessionId:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        expiresAt:
          type: string
          format: date-time
    CompleteQuizSessionRequest:
      type: object
      required: [answers, idempotencyKey]
      properties:
        answers:
          type: array
          items:
            type: integer
            nullable: true
          description: Array of selected option indices (null = unanswered). Length must equal totalQuestions.
        idempotencyKey:
          type: string
          minLength: 1
    QuizAttempt:
      type: object
      required: [id, userId, materialId, score, totalQuestions, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        materialId:
          type: string
        score:
          type: integer
        totalQuestions:
          type: integer
        earnedCoins:
          type: integer
        createdAt:
          type: string
          format: date-time
    FlashcardsRequest:
      type: object
      required: [materialIds]
      properties:
        materialIds:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 20
          description: Owned, study-ready material ids.
    Flashcard:
      type: object
      required: [front, back, materialTitle]
      properties:
        front:
          type: string
        back:
          type: string
        materialTitle:
          type: string
    FlashcardsResponse:
      type: object
      required: [flashcards]
      properties:
        flashcards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
...