# lib/api-zod/src/generated/types/startQuizSessionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StartQuizSessionRequestTotalQuestions } from './startQuizSessionRequestTotalQuestions';

export interface StartQuizSessionRequest {
  /**
     * Owned material ids to combine into one unified quiz.
     * @minItems 1
     * @maxItems 20
     */
  materialIds: string[];
  totalQuestions: StartQuizSessionRequestTotalQuestions;
}